pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadence-python-client"
version = "0.1.0"
description = "Python framework for authoring Cadence workflows and activities"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Cadence"}
]
keywords = ["workflow", "orchestration", "distributed", "async", "cadence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "grpcio==1.71.2",
    "grpcio-status>=1.71.2",
    "msgspec>=0.19.0",
    "protobuf==5.29.1",
    "typing-extensions>=4.0.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools==1.71.2",
    "pytest>=8.4.1",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pytest-docker>=3.2.3",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]
examples = [
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
]

[project.urls]
Homepage = "https://cadenceworkflow.io/"
Documentation = "https://cadenceworkflow.io/docs/get-started"
Repository = "https://github.com/cadence-workflow/cadence-python-client"
"Bug Tracker" = "https://github.com/cadence-workflow/cadence-python-client/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cadence*"]

[tool.setuptools.package-data]
"*" = ["*.proto"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Temporarily disable to allow CI to pass
disallow_incomplete_defs = false  # Temporarily disable to allow CI to pass
check_untyped_defs = true
disallow_untyped_decorators = false  # Temporarily disable to allow CI to pass
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
explicit_package_bases = true
exclude = [
    "cadence/api",
    "cadence/api/.*",
    "cadence/sample",
]

# Reduce recursive module checking
follow_imports = "silent"

[[tool.mypy.overrides]]
module = [
    "grpcio.*",
    "grpcio_tools.*",
    "grpc.*",
    "google.protobuf.*",
    "uber.cadence.*",
    "msgspec.*",
    "cadence.api.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config --import-mode=importlib"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "asyncio: marks tests as async tests",
]

[tool.coverage.run]
source = ["cadence"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc.)?abstractmethod",
]

[tool.ruff]
exclude = [
    "cadence/api", # generated proto files
]
